+++ index.html
@@ -4,9 +4,9 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Virtual Keyboard</title>
-  <link rel="stylesheet" href="./style.css">
+  <link rel="stylesheet" href="./src/styles/style.css">
 </head>
 <body>
-  <script src="./script.js"></script>
+  <script src="./dist/main.js"></script>
 </body>
 </html>
